+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - String only dropdown button implementation.
+## [0.0.3] - Any type dropdown button implementation.
+
+* Create dropdown button for any data items.
+
+## [0.0.2] - String only dropdown button implementation.
+
+* Create dropdown button for string items easily.
+
+## [0.0.1] - Simple dropdown button for string values.
 
 * Create dropdown button for string items easily.